Conflicts:
	application/controllers/admin/index.html
	application/models/form_m.php
	assets/site/style/quickform.css

	new file:   application/controllers/admin/index.html
	new file:   application/libraries/search_db.php
	modified:   assets/site/style/css.css
	modified:   assets/site/style/quickform.css
	modified:   guinnessapp.sql

+++ application/models/app_m.php
@@ -51,15 +51,13 @@ Class App_m extends CI_model {
 	 
 	public function remove($gid)
 	{
-		$deleted = $this->db->query("DELETE FROM campaign_app WHERE APP_APPLICATION_ID = ".$gid);
-
-		if(!$deleted){
-		$this->error[] = "Deleting has Failed, Try Again or Contact Web Administrator";
+	  $this->load->model('campaign_m');
+	  if($campaign = $this->campaign_m->getCampaignByAppID($gid,true)){
 		return false;
-		}else{
-		return true;
-		}
-		
+	  }else{
+		$deleted = $this->db->query("DELETE FROM campaign_app WHERE APP_APPLICATION_ID = ".$gid);
+		return $deleted ? true : false;
+	  }
 	}	 
 	  
 	public function retrieve($clauses = array() , $args = array())
